use CGFloat, instead of Float

+++ SABlurImageView/UIImage+BlurEffect.swift
@@ -15,17 +15,17 @@ extension UIImage {
         return UIImage(CGImage: cgImage)
     }
     
-    func blurEffect(boxSize: Float) -> UIImage! {
+    func blurEffect(boxSize: CGFloat) -> UIImage! {
         return UIImage(CGImage: bluredCGImage(boxSize))
     }
     
-    func bluredCGImage(boxSize: Float) -> CGImageRef! {
+    func bluredCGImage(boxSize: CGFloat) -> CGImageRef! {
         return CGImage!.blurEffect(boxSize)
     }
 }
 
 extension CGImage {
-    func blurEffect(boxSize: Float) -> CGImageRef! {
+    func blurEffect(boxSize: CGFloat) -> CGImageRef! {
         
         let boxSize = boxSize - (boxSize % 2) + 1
